Added DevProjects under JavaScript challenges, ReactJS, and Vue. Removed JavaScript Gekk from JavaScript challenges because there's no link to it and I couldn't find it on Google.Added Arc.dev to job sites. Added Codementor to freelance jobs sites.
DevProjects is a community with free real-world projects in multiple coding languages and tech stacks. The website also includes solutions from other developers and discussion board for all users to participate in.

Arc.dev crawls the web to find every remote job and gives developers a personalized list of jobs they are more likely to land.

Codementor is a platform where developers can join to provide others mentorship or look for freelance jobs.